+++ .eslintrc.js
@@ -1,0 +1,54 @@
+module.exports = {
+  env: {
+    browser: true,
+    es2020: true,
+    node: true,
+    'jest/globals': true,
+  },
+  extends: [
+    'airbnb-base',
+    'plugin:jest/recommended',
+    'plugin:@typescript-eslint/eslint-recommended',
+    'plugin:@typescript-eslint/recommended',
+    'plugin:@typescript-eslint/recommended-requiring-type-checking',
+  ],
+  parser: '@typescript-eslint/parser',
+  parserOptions: {
+    ecmaVersion: 11,
+    sourceType: 'module',
+    project: './tsconfig.eslint.json',
+    tsconfigRootDir: __dirname,
+  },
+  plugins: [
+    'jest',
+    '@typescript-eslint',
+  ],
+  rules: {
+    semi: ['error', 'never'],
+    'class-methods-use-this': ['error', { exceptMethods: ['fold', 'escape'] }],
+    'import/extensions': [
+      'error',
+      'ignorePackages',
+      {
+        js: 'never',
+        ts: 'never',
+      },
+    ],
+  },
+  overrides: [
+    {
+      files: ['lib/**/__tests__/*'],
+      rules: {
+        '@typescript-eslint/no-unsafe-member-access': 'off',
+        '@typescript-eslint/no-unsafe-call': 'off',
+      },
+    },
+  ],
+  settings: {
+    'import/resolver': {
+      node: {
+        extensions: ['.js', '.ts'],
+      },
+    },
+  },
+}