+++ lib/types/VCard.ts
@@ -1,0 +1,8 @@
+export interface Property {
+  key: string
+  value: string
+}
+
+export interface DefinedElements {
+  [key: string]: boolean
+}